requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
pypdf
tiktoken
sentence-transformers
